changed corpus to use long document ids, tests pass, but have possibly introduced some errors anyway, these will be found as the upper interfaces are modified to use longs

+++ core/src/test/java/org/lemurproject/galago/core/index/corpus/DocumentToKeyValuePairTest.java
@@ -46,7 +46,7 @@ public class DocumentToKeyValuePairTest extends TestCase {
     dkvp.process(document);
 
     KeyValuePair pair = kvpProcessor.pair;
-    assertEquals(Utility.toInt(pair.key), 1);
+    assertEquals(Utility.toLong(pair.key), 1);
 
     ByteArrayInputStream stream = new ByteArrayInputStream(pair.value);
     ObjectInputStream input = new ObjectInputStream(stream);
